requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.3
backports-abc==0.5
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
cx-Freeze==5.1.1
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
dis3==0.1.2
Django==1.11.16
docutils==0.14
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
enum34==1.1.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gevent==1.3.7
google-search-results==1.2.4
greenlet==0.4.15
HTMLParser==0.0.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
Kivy==1.10.1
Kivy-examples==1.10.1
Kivy-Garden==0.1.4
kivy.deps.glew==0.1.10
kivy.deps.gstreamer==0.1.13
kivy.deps.sdl2==0.1.18
lxml==4.2.5
macholib==1.11
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
numpy==1.15.4
parsel==1.5.1
pathlib==1.0.1
pefile==2018.8.8
plac==0.9.6
preshed==2.0.1
psutil==5.4.8
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
PyDispatcher==2.0.5
pygame==1.9.4
Pygments==2.2.0
PyHamcrest==1.9.0
PyInstaller==3.4
pyOpenSSL==18.0.0
pypiwin32==223
pyreadline==2.1
pyserial==3.4
PySide==1.2.4
pytz==2018.7
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==3.13
QDarkStyle==2.6.4
qtgui==0.0.1
queuelib==1.5.0
regex==2018.1.10
requests==2.20.1
Scrapy==1.5.1
service-identity==17.0.0
singledispatch==3.4.0.3
six==1.11.0
spacy==2.0.18
SQLAlchemy==1.2.14
thinc==6.12.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
Twisted==18.9.0
ujson==1.35
urllib3==1.24.1
w3lib==1.19.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.6.0
